Automate PDF generation of tabular data
This is an initial implementation. Only handling
String data for now, also using only default styling.
Also not handling cellwidth for now, just setting
to wrap text for all cells.

Anyone wanting to use automatically generate PDF of tabular
data should:

1. It is assumed that the data is in the form of a list of a rows. So,
implement RowMapper to specify the alignment of data of a particular row.

2. Construct a TabularDocumentDefinition object, which contains/wraps all
the information needed to generate a table. The no. of columns, the
column headers, the list of the data, and the aformentioned rowmapper.

3. Construct a TabularDocBuilder using the above tabularDocumentDefinition and
call the createDocument() method. PDF will be generated.