Automate PDF generation of a business letter
This is an initial implementation. This generates a letter that conforms to
the business letter format specified in https://owl.english.purdue.edu/owl/resource/653/01.

Anyone wanting to use automatically generate PDF of business letter should
construct a LetterDocumentDefinition object, which contains/wraps all
the information needed to generate a business letter.